refactor: rename childProcess -> processUtils
Another PR I have adds a process related function, so I thought I'd make
the name more appropriate
