+++ packages/core/src/shared/remoteSession.ts
@@ -17,7 +17,7 @@ import { ToolkitError, UnknownError } from './errors'
 import { getLogger } from './logger/logger'
 import { getOrInstallCli } from './utilities/cliUtils'
 import { pushIf } from './utilities/collectionUtils'
-import { ChildProcess } from './utilities/childProcess'
+import { ChildProcess } from './utilities/processUtils'
 import { findSshPath, getVscodeCliPath } from './utilities/pathFind'
 import { IamClient } from './clients/iamClient'
 import { IAM } from 'aws-sdk'
