+++ packages/core/src/shared/sam/debugger/csharpSamDebug.ts
@@ -18,7 +18,7 @@ import { ExtContext } from '../../extensions'
 import { DefaultSamLocalInvokeCommand, waitForDebuggerMessages } from '../cli/samCliLocalInvoke'
 import { runLambdaFunction, waitForPort } from '../localLambdaRunner'
 import { SamLaunchRequestArgs } from './awsSamDebugger'
-import { ChildProcess } from '../../utilities/childProcess'
+import { ChildProcess } from '../../utilities/processUtils'
 import { HttpResourceFetcher } from '../../resourcefetcher/httpResourceFetcher'
 import { getLogger } from '../../logger'
 import * as vscode from 'vscode'
