+++ packages/core/src/shared/sam/debugger/goSamDebug.ts
@@ -17,7 +17,7 @@ import { SamLaunchRequestArgs } from './awsSamDebugger'
 import { getLogger } from '../../logger'
 import * as fs from 'fs-extra'
 import fs2 from '../../fs/fs'
-import { ChildProcess } from '../../utilities/childProcess'
+import { ChildProcess } from '../../utilities/processUtils'
 import { Timeout } from '../../utilities/timeoutUtils'
 import { execFileSync, SpawnOptions } from 'child_process'
 import * as nls from 'vscode-nls'
