+++ packages/core/src/shared/sam/debugger/typescriptSamDebug.ts
@@ -9,7 +9,7 @@ import * as path from 'path'
 import * as vscode from 'vscode'
 import { isImageLambdaConfig, NodejsDebugConfiguration } from '../../../lambda/local/debugConfiguration'
 import { RuntimeFamily } from '../../../lambda/models/samLambdaRuntime'
-import { ChildProcess } from '../../../shared/utilities/childProcess'
+import { ChildProcess } from '../../utilities/processUtils'
 import * as pathutil from '../../../shared/utilities/pathUtils'
 import { ExtContext } from '../../extensions'
 import { getLogger } from '../../logger'
