+++ packages/core/src/shared/sam/localLambdaRunner.ts
@@ -24,7 +24,7 @@ import { asEnvironmentVariables } from '../../auth/credentials/utils'
 import { buildSamCliStartApiArguments } from './cli/samCliStartApi'
 import { DefaultSamCliProcessInvoker } from './cli/samCliInvoker'
 import { APIGatewayProperties } from './debugger/awsSamDebugConfiguration.gen'
-import { ChildProcess } from '../utilities/childProcess'
+import { ChildProcess } from '../utilities/processUtils'
 import { SamCliSettings } from './cli/samCliSettings'
 import * as CloudFormation from '../cloudformation/cloudformation'
 import { sleep } from '../utilities/timeoutUtils'
