+++ packages/core/src/shared/sam/sync.ts
@@ -17,7 +17,7 @@ import * as CloudFormation from '../cloudformation/cloudformation'
 import { DefaultEcrClient } from '../clients/ecrClient'
 import { createRegionPrompter } from '../ui/common/region'
 import { CancellationError } from '../utilities/timeoutUtils'
-import { ChildProcess, ChildProcessResult } from '../utilities/childProcess'
+import { ChildProcess, ChildProcessResult } from '../utilities/processUtils'
 import { keys, selectFrom } from '../utilities/tsUtils'
 import { Commands } from '../vscode/commands2'
 import { AWSTreeNodeBase } from '../treeview/nodes/awsTreeNodeBase'
