+++ packages/core/src/shared/sshConfig.ts
@@ -8,7 +8,7 @@ import * as path from 'path'
 import * as fs from 'fs-extra'
 import * as nls from 'vscode-nls'
 import { getLogger } from './logger'
-import { ChildProcess, ChildProcessResult } from './utilities/childProcess'
+import { ChildProcess, ChildProcessResult } from './utilities/processUtils'
 import { Result } from './utilities/result'
 import { ToolkitError } from './errors'
 import { getIdeProperties } from './extensionUtilities'
