+++ packages/core/src/test/awsService/ec2/sshKeyPair.test.ts
@@ -9,7 +9,7 @@ import * as path from 'path'
 import { SshKeyPair } from '../../../awsService/ec2/sshKeyPair'
 import { createTestWorkspaceFolder, installFakeClock } from '../../testUtil'
 import { InstalledClock } from '@sinonjs/fake-timers'
-import { ChildProcess } from '../../../shared/utilities/childProcess'
+import { ChildProcess } from '../../../shared/utilities/processUtils'
 import { fs } from '../../../shared'
 
 describe('SshKeyUtility', async function () {
