+++ packages/core/src/test/shared/codelens/localLambdaRunner.test.ts
@@ -9,7 +9,7 @@ import * as fs from 'fs-extra'
 import * as fsUtils from '../../../shared/filesystemUtilities'
 import { SamCliBuildInvocation, SamCliBuildInvocationArguments } from '../../../shared/sam/cli/samCliBuild'
 import * as localLambdaRunner from '../../../shared/sam/localLambdaRunner'
-import { ChildProcessResult } from '../../../shared/utilities/childProcess'
+import { ChildProcessResult } from '../../../shared/utilities/processUtils'
 import { FakeExtensionContext } from '../../fakeExtensionContext'
 import { SamLaunchRequestArgs } from '../../../shared/sam/debugger/awsSamDebugger'
 import { assertLogsContain } from '../../globalSetup.test'
