+++ packages/core/src/test/shared/utilities/cliUtils.test.ts
@@ -8,7 +8,7 @@ import * as fs from 'fs-extra'
 import * as path from 'path'
 import { installCli } from '../../../shared/utilities/cliUtils'
 import globals from '../../../shared/extensionGlobals'
-import { ChildProcess } from '../../../shared/utilities/childProcess'
+import { ChildProcess } from '../../../shared/utilities/processUtils'
 import { SeverityLevel } from '../vscode/message'
 import { assertTelemetryCurried } from '../../testUtil'
 import { getTestWindow } from '../../shared/vscode/window'
